API1/requirements.txt
amqp==5.1.1
asgiref==3.7.2
asyncio==3.4.3
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
django-extensions==3.2.3
djangorestframework==3.14.0
idna==3.4
kombu==5.3.1
prompt-toolkit==3.0.38
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
